More refactoring of CUE
