+++ src/main/java/com/vesperin/cue/cmds/TypicalityAnalysisCommand.java
@@ -7,6 +7,7 @@ import com.google.common.base.Strings;
 import com.vesperin.base.Source;
 import com.vesperin.cue.Cue;
 import com.vesperin.cue.Introspector;
+import com.vesperin.cue.IntrospectorWithCli;
 import com.vesperin.cue.utils.IO;
 import com.vesperin.cue.utils.Sources;
 
@@ -22,8 +23,8 @@ import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-import static com.vesperin.cue.cmds.CallableCommand.allNonNull;
-import static com.vesperin.cue.cmds.CallableCommand.allNull;
+import static com.vesperin.cue.IntrospectorWithCli.allNonNull;
+import static com.vesperin.cue.IntrospectorWithCli.allNull;
 import static java.nio.file.StandardOpenOption.APPEND;
 import static java.nio.file.StandardOpenOption.CREATE;
 
@@ -31,7 +32,7 @@ import static java.nio.file.StandardOpenOption.CREATE;
  * @author Huascar Sanchez
  */
 @Command(name = "typical", description = "Finds the top k most typical implementation")
-public class TypicalityAnalysisCommand implements CallableCommand {
+public class TypicalityAnalysisCommand implements IntrospectorWithCli.CliCommand {
 
   private static final String TYPICAL_SET_FILE_NAME = "typicalset.txt";
